refactor: relocate telemetry and utils to their common homes
pkg/telemetry.Tracer was a verbatim duplicate of
pkg/common/observability/tracing.Tracer (the callers all pass an explicit
scope, so the differing default name was never used); drop the package and
point the kvcache tracing call sites at the observability package. Align
their instrumentation scope strings with the router identity
(llm-d-router/pkg/...).

Move the generic SliceMap helper from pkg/utils to a new
pkg/common/collections package. Fast follow to the kvcache/kvevents
internalization (#1886).
