+++ .golangci.yml
@@ -46,7 +46,7 @@ linters:
       - linters: [revive]
         text: "avoid package names that conflict with Go standard library package names"
       # Packages transferred verbatim from llm-d-kv-cache (#1862); style cleanup tracked separately
-      - path: pkg/(kvcache|kvevents|telemetry|utils)/
+      - path: pkg/(kvcache|kvevents)/
         linters:
           - revive
           - perfsprint
