+++ pkg/common/collections/README.md
@@ -1,0 +1,8 @@
+# Collections
+
+Generic helpers for working with slices and maps that the standard library does
+not provide. Small, dependency-free utilities shared across the router.
+
+| Symbol | Role |
+|--------|------|
+| `SliceMap` | Apply a function to each element of a slice, returning a new slice. |
